add queries support, convert schema to queries
