+++ packages/slate-prop-types/src/index.js
@@ -8,7 +8,6 @@ import {
   Mark,
   Node,
   Range,
-  Schema,
   Value,
   Text,
 } from 'slate'
@@ -68,7 +67,6 @@ const Types = {
   nodes: create('List<Node>', v => Node.isNodeList(v)),
   range: create('Range', v => Range.isRange(v)),
   ranges: create('List<Range>', v => Range.isRangeList(v)),
-  schema: create('Schema', v => Schema.isSchema(v)),
   value: create('Value', v => Value.isValue(v)),
   text: create('Text', v => Text.isText(v)),
   texts: create('List<Text>', v => Text.isTextList(v)),
