+++ packages/slate-react/src/components/leaf.js
@@ -126,11 +126,10 @@ class Leaf extends React.Component {
 
   renderText() {
     const { block, node, editor, parent, text, index, leaves } = this.props
-    const { schema } = editor
 
     // COMPAT: Render text inside void nodes with a zero-width space.
     // So the node can contain selection but the text is not visible.
-    if (schema.isVoid(parent)) {
+    if (editor.query('isVoid', parent)) {
       return <span data-slate-zero-width="z">{'\u200B'}</span>
     }
