+++ packages/slate/src/utils/is-object.js
@@ -10,16 +10,13 @@ export const TYPES = {
   decoration: '@@__SLATE_DECORATION__@@',
   document: '@@__SLATE_DOCUMENT__@@',
   editor: '@@__SLATE_EDITOR__@@',
-  history: '@@__SLATE_HISTORY__@@',
   inline: '@@__SLATE_INLINE__@@',
   leaf: '@@__SLATE_LEAF__@@',
   mark: '@@__SLATE_MARK__@@',
   operation: '@@__SLATE_OPERATION__@@',
   point: '@@__SLATE_POINT__@@',
   range: '@@__SLATE_RANGE__@@',
-  schema: '@@__SLATE_SCHEMA__@@',
   selection: '@@__SLATE_SELECTION__@@',
-  stack: '@@__SLATE_STACK__@@',
   text: '@@__SLATE_TEXT__@@',
   value: '@@__SLATE_VALUE__@@',
 }
